add: author and genre dao with tests

+++ hw-06/src/main/java/ru/otus/booklibrary/repo/AuthorDao.java
@@ -4,7 +4,7 @@ import ru.otus.booklibrary.domain.Author;
 
 public interface AuthorDao {
 
-    Integer count();
+    Long count();
 
     Author getByName(String name);
